move interface ot usecase
